Make the user interface allow getUser to pick which fields to return
It's now possible to specify which fields of the user should be returned
by getUser when called on the User model. This allows model designers to
limit the amount of data transferred between a remote datastore and the
local server. Instead of getting the whole user everytime, the request
now only retrieves and returns a subset of the user.
If no selection parameter is given, the whole user is still returned.